seed() now accesses the "leaf seed" of a DelayedArray object
If 'x' is a DelayedArray object and the tree of DelayedOp objects stored
in 'x@seed' is linear, then seed(x) must access the "leaf seed" i.e. the
1st node in the linear tree (starting from 'x@seed', the tree root) that
is not a DelayedOp object. This should be the seed that was used by the
user to construct the original DelayedArray object.

Both the seed() getter and setter now raise an error if the tree in 'x'
is not linear.