"[" method for DelayedArray objects now honors the 'drop' argument
When used in "multi-dimensional mode", that is, when called with one
(possibly missing) subscript per dimension (e.g. 'x[6:2, , 5]'), the
"[" method for DelayedArray objects used to behave like if the 'drop'
argument was always set to FALSE (hence ignoring the value passed to it).
Now it honors the 'drop' argument. This means that "ineffective"
dimensions (i.e. dimensions equal to 1) are now dropped by default.

For example, if 'x' is a 6x10x8 DelayedArray object, then 'x[-1, , 5]'
now returns a DelayedMatrix object and 'x[1, , 5]' a mono-dimensional
DelayedArray object.