DelayedArray now extends DelayedUnaryOp
In the tree of DelayedOp objects stored in the "seed" slot of a
DelayedArray object, the DelayedArray object itself can be seen as
the root node and this node can be seen as representing a no-op.
Making DelayedArray a subclass of DelayedOp formalizes this and allows
some significant code simplifications. In particular any code that
needs to walk the tree does not need to treat the root node any
differently than the other internal nodes.