Add a "bindROWS" method for DelayedArray objects
This "bindROWS" method for DelayedArray objects does arbind(). This
makes rbind() work on DataFrame objects with DelayedArray objects in
their columns (including when the columns are 1D DelayedArray objects).

acbind() now fails graciously on 1D arrays or DelayedArray objects.

rbind() and cbind() now are supported on DelayedArray objects in general,
not just on DelayedMatrix objects (as long as the objects to bind have at
least 2 dimensions for cbind()), and are equivalent to arbind() and acbind(),
respectively.