No-op operations on a DelayedArray derivative really act like no-ops
Operating on a DelayedArray derivative (e.g. RleArray, HDF5Array or
GDSArray) will now return an objet of the original class if the result
is "pristine" (i.e. if it doesn't carry delayed operations) instead of
degrading the object to a DelayedArray instance. This applies for example
to 't(t(x))' or 'dimnames(x) <- dimnames(x)' etc...

Note that this doesn't apply to operations that are conceptually no-ops
but don't return a pristine object e.g. x + 0 still degrades to a
DelayedArray instance because this operation gets stashed in x (as a
DelayedOp object) despite being conceptually a no-op. This is due to
the current (but intended) limitation of the no-op detection code
(performed by isNoOp()).