Some changes to makeCappedVolumeBox() and defaultGrid()
The 'shape' argument of makeCappedVolumeBox(), 'viewport_shape' argument
of makeRegularArrayGridOfCappedLengthViewports(), and 'block.shape'
argument of defaultGrid() now can be set to "hypercube", "proportional",
"first-dim-grows-first", or "last-dim-grows-first".

The "first-dim-grows-first" shape replaces the "linear" shape (same shape
but new name).

The "last-dim-grows-first" shape is new.

The "proportional" shape is not implemented yet (just a placeholder for
now).