Fix another performance regression introduced at commit e1cee763
Commit e1cee763 introduced another severe preformance regression to the
"show" method for DelayedArray objects. This time the regression was
occuring in the context of displaying objects of 3 dimensions or more.
In order to display 2D slices of the object, the slices were fully loaded
into memory.
This fix uses [ instead of extract_array() to obtain the 2D slices so the
slices are no more loaded in memory ([ is delayed). As a result only the
elements of the 2D slices that need to be displayed get loaded into memory.