Proposed improvements to ChunkedRleArraySeed objects
Add some proposed modifications (in the form of commented code for now so
not effective) to the definitions and validity methods of RleRealizationSink
and ChunkedRleArraySeed objects. The purpose of these modifications is to
make these objects support a RegularArrayGrid of chunks, a long due
improvement that still needs to happen.