Workers now inherit DelayedArray user-controlled global options
In the context of BiocParallel::bplapply() and family, workers now inherit
the user-controlled options defined on the master. Workers can also modify
the inherited options or define new user-controlled options but this does
not affect the options defined on the master (one-way communication).
This mechanism works with nested levels of parallelization i.e. if workers
start subworkers and so on.