Reduce memory usage when realizing a sparse DelayedArray to disk
On-disk realization of a DelayedArray object that is reported to be sparse
(by is_sparse()) to a "sparsity-optimized" backend (i.e. to a backend with
a memory efficient write_sparse_block() like the TENxMatrix backend imple-
mented in the HDF5Array package) now preserves sparse representation of the
data all the way. More precisely, each block of data is now kept in a sparse
form during the 3 steps that it goes thru: read from seed, realize in
memory, and write to disk.