Multi-dimensional subsetting now returns 1D array as ordinary vector
Multi-dimensional single bracket subsetting of a DelayedArray object x
(e.g. x[3:1, , 6]) was always delayed until now, which means that the
result was returned as another DelayedArray object carrying the subsetting
operation as a delayed operation. There is now one exception to this rule:
when drop=TRUE and the result has only one dimension, then it is returned
as an **ordinary** vector (atomic or list). This means that the subsetting
operation is no more delayed but realilzed in memory.

See https://github.com/Bioconductor/DelayedArray/issues/6 for the
motivation behind this important change.