Improve RleArray() contructor
In addition to an Rle object, RleArray() now accepts a list-like object
where all the list elements are Rle objects. This can be an ordinary list
of Rle objects, an RleList object, a DataFrame object where all the columns
are Rle objects, etc... This makes it easier to construct a large RleArray
object (i.e. an object of length >= 2^31).

Other improvements to RleArray():
  - Improve error message when the supplied Rle object is too long (i.e.
    has a length >= 2^31).
  - Check that the length of the input data matches the supplied dimensions.

Other changes:
  - Coercion from RleList to RleMatrix now supports RleList objects that
    have a total length >= 2^31.
  - Coercion from RleMatrix to RleList or DataFrame now supports RleMatrix
    objects that have a length >= 2^31.