More improvements to RleArray() contructor
If the supplied 'data' is a list-like object where all the list elements
are Rle objects of the same length, and if the 'dim' argument is not
specified, then the RleArray() constructor now works and returns an
RleMatrix object with 1 column per Rle object in 'data'. If the 'dimnames'
argument is not specified, then the names on 'data' are propagated as
the colnames of the returned object.