Yet another tweak to the dimnames() setter for DelayedArray objects
The supplied dimnames now can be a "truncated list" i.e. the list can be
shorter than the number of dimensions. This is needed in order to support
the following alternative way of setting rownames on a DelayedArray
object 'x':

    dimnames(x)[1] <- list(somenames)

If 'x' has no dimnames yet (i.e. if 'dimnames(x)' is NULL) then `dimnames<-`
gets called with a right value that is a list of length 1. Note that the
above code is actually how the latest version of utils::tail.matrix() sets
the rownames on 'x' when 'keepnums' is TRUE (this is a recent change in
R 4.0). And because utils::tail.matrix() gets called when doing tail()
on a DelayedMatrix object 'x', 'tail(x)' was failing if 'x' didn't have
dimnames:

    Error in normarg_dimnames(dimnames, seed_dim) :
      the supplied 'dimnames' must have one list element per dimension