Support 'type(x) <- new_type' to change the type of a DelayedArray object
'type(x) <- new_type' is implemented as a delayed operation.

See new example in '?RleArray' (E. SWITCH THE TYPE OF AN RleArray
OBJECT FROM "double" TO "integer") for a commented example of how
to **effectively** switch the type of an RleArray object from "double"
to "integer" in order to reduce its memory footprint.