blockApply() and blockReduce() no longer set attributes on the blocks
The purpose of these attributes was to provide the grid context of
the current block. Unfortunately setting these attributes can trigger
a copy of the block (e.g. when the block is an S4 object).
See https://github.com/Bioconductor/DelayedArray/issues/69

So blockApply() and blockReduce() no longer set these attributes and the
little helper functions that the user was expected to use to retrieve these
attributes (effectiveGrid(), currentBlockId(), and currentViewport()) now
use an entirely different mechanism to return the grid context.

Note that the block attributes were not documented because not expected
to be accessed directly so this change shouldn't affect existing code.