Make sure updateObject() works on old DelayedArray objects even when they contain old HDF5ArraySeed instances that point to non-existing files.
For example ExperimentHub resource EH1656 is a serialized
SummarizedExperiment object where the assay is such a difficult
DelayedArray object. The object points to non-existing file 'assays.h5'
so is broken since its addition to ExperimentHub on 2017-10-06 by the
restfulSEData folks. However, we still want to be able to call
updateObject( , check=FALSE) on it. Note that doing so updates its
internal representation but doesn't fix the main issue that the
object is pointing to a non-existing file.