More improvements to BLOCK_rowsum() and BLOCK_colsum()
When 'BACKEND' is not NULL (e.g. is set to "HDF5Array" or "TENxMatrix"),
and 'BPPARAM' is NULL (i.e. grid strips are processed sequentially), then
most of the time we are now able to return a **pristine** DelayedMatrix
object instead of one made of a bunch of pristine DelayedMatrix objects
bound together (via rbind() or cbind()).
Situations where we cannot do it is when the internal walk on the input
matrix is incompatible with the chunking of the realization sink. This
happens for example with:

  m <- matrix(runif(60), nrow=5)
  group <- sample(letters[1:3], ncol(m), replace=TRUE)
  BLOCK_colsum(m, group, grid=rowAutoGrid(m, 2), BACKEND="TENxMatrix")

Here the returned object is a DelayedMatrix object obtained by rbind'ing
3 TENxMatrix objects together, instead of a single TENxMatrix object.