Improvement to BLOCK_mult_Lgridi() and BLOCK_mult_Rgrid()
Internal functions BLOCK_mult_Lgrid() and BLOCK_mult_Rgrid() are the
workhorses behind matrix multiplication between a DelayedMatrix object
(or derivative) and a non-DelayedMatrix object.

This change implements the same improvement that was implemented in
BLOCK_rowsum() and BLOCK_colsum() at commit 1f6eeb4.

When 'BACKEND' is not NULL (e.g. is set to "HDF5Array" or "TENxMatrix"),
and 'BPPARAM' is NULL (i.e. grid strips are processed sequentially), then
most of the time we are now able to return a **pristine** DelayedMatrix
object instead of one made of a bunch of pristine DelayedMatrix objects
bound together (via rbind() or cbind()).
Situations where we cannot do it is when the internal walk on the input
matrix is incompatible with the chunking of the realization sink. This
happens for example with:

  m1 <- matrix(1:12, ncol=3)
  m2 <- matrix(101:115, nrow=3)
  BLOCK_mult_Lgrid(m1, m2, Lgrid=rowAutoGrid(m1, 2), BACKEND="TENxMatrix")

Here the returned object is a DelayedMatrix object obtained by rbind'ing
2 TENxMatrix objects together, instead of a single TENxMatrix object.