More improvements to BLOCK_rowsum/colsum and BLOCK_mult_Lgrid/Rgrid
Add 'dry.run' and ellipsis arguments.

When 'dry.run' is set to TRUE, a succinct description of what the result
would look like is returned in a named list. If 'BACKEND' is NULL, this
list contains components "class", "dim", and "type". Otherwise, it
contains the additional "nseed" component which is either set to 1 if
the result is expected to be a pristine DelayedMatrix object, or to
the number of pristine DelayedMatrix objects bound together. Note that
the "nseed" value is what the nseed() function will actually produce on the
DelayedMatrix object effectively returned when 'dry.run' is set to FALSE.

Arguments passed thru the ellipsis are passed to the internal call to the
RealizationSink constructor associated with the specified backend. This
only applies when 'BACKEND' is not NULL and the returned DelayedMatrix
object is expected to be pristine (i.e. nseed == 1). Note that the first
3 arguments of **any** RealizationSink constructor are guaranteed to
be 'dim', 'dimnames', and 'type', and the arguments specified thru
the ellipsis of BLOCK_rowsum/colsum or BLOCK_mult_Lgrid/Rgrid can not be
any of these. Also 'as.sparse' -- which is supported by some RealizationSink
constructors but not all -- is not allowed either.